Hdsk.Bayesian.Bins

Description

Documentation

bin :: (Functor f, Selectable f) => [Double -> Bool] -> f Double -> f Int

O(n k^2 n log n) where k is # of bins. Map each number in a vector of continueous data to a bin number according to its percentile (makes even-width bins). Bins are 0-indexed. Based on sample quantiles.

genIntervals :: Selectable f => Int -> f Double -> [Double -> Bool]

O(kn) Generate the bounds of fixed width bins, encoded as predicates of sample quantiles.