Hdsk.Bayesian.NaiveBayes

Description

A simple implementation of a Naive Bayes classifier for purely categorical data.

Documentation

classify :: [Tuple] -> Tuple -> String

Calculate the most probable label of the given tuple given the dataset (list of tuples). The label of a data object should be the last field.