Hdsk.Cluster

Description

This module provides an implementation of the k-means clustering algorithm. K-means is well suited to problems where the number of clusters is known and fixed (it is a parameter to the algorithm).

Documentation

kmeans :: Int -> [[Double]] -> [Int]

O(inkD) Run the kmeans clustering algorithm over the given dataset. Returns a list of cluster labels which corresponds 1-1 with the input list of data points.

kmedoids :: Int -> [[Double]] -> [Int]

O(ikD*n^2) Run the kmedoids clustering algorithm over the given dataset.

kclusterer

Arguments

:: Eq tup

=> Double
Minimum improvement eta

-> (tup -> tup -> Double)
Distance metric between points

-> ([tup] -> tup)
Measure of center of cluster

-> Int
The parameter k

-> [tup]
The list of data points

-> [Int]
Final clustering

Convenience function for creating clustering function. For instance, kmeans, as defined in this module, is a kclusterer with distEuclidean distance metric, centroid center measure, and eta = 0.01.

cluster :: (Eq tup, Ord d) => (tup -> tup -> d) -> [tup] -> [tup] -> [Int]

O(nkD) where n is the number of data points, k is the number of clusters, and D is the dimensionality of the data. Run one iteration of the k-means algorithm. The parameter k, number of clusters, is implied by the length of the list of initial centroids. Returns a list of cluster labels (encoded as integers) which correspond 1-1 with the given list of data points.

improve :: (Eq tup, Ord d) => (tup -> tup -> d) -> ([tup] -> tup) -> [tup] -> [Int] -> [[Int]]

O(inkD) where n is the number of data points, k is the number of clusters, D is the dimensionality of the data, and i is the number of iterations. Improve the quality of the clustering. Generates an infinite list of clusterings.

meanSqDist :: (Ord d, Floating d) => (tup -> tup -> d) -> ([tup] -> tup) -> [tup] -> [Int] -> d

O(nD) Calculate the mean squared distance from each point in a cluster to the centroid.

midpoints :: ([tup] -> tup) -> [Int] -> [tup] -> [tup]

O(nD) where n is the number of data points and D is the dimensionality of the data. Calculate the midpoints of clusters according to a metric (centroid, for instance).

centroid :: Fractional a => [[a]] -> [a]

O(nD) where n is the number of points and D is the dimensionality of each point. Compute the mean of a list of D-dimensional points.

medoid :: (Ord d, Fractional d) => (tup -> tup -> d) -> [tup] -> tup

O(n^2) Compute the medoid of a list of points.

closestTo :: Ord d => (tup -> tup -> d) -> tup -> [tup] -> tup

O(kD) where k is the number of points to consider and D is the dimensionality of the data. Select from a list of points that which is closest to a given point x according to a given distance metric. For instance, in one dimension:

>>> closestTo (\x y -> abs $ sum (zipWith (-) x y)) [0] [[1], [2]]
>>> [1.0]

minkowski :: Floating a => a -> [a] -> [a] -> a

O(D) where D is the dimensionality of the vectors. General distance metric between two vectors/ data points in the same vector space. The first parameter p corresponds to the L^p space to compute the norm (e.g. p = 2 is Euclidean distance).

distManhattan :: Floating a => [a] -> [a] -> a

Common case of minkowski is with p = 1 for Manhattan distance.

distEuclidean :: Floating a => [a] -> [a] -> a

Common case of minkowski is with p = 2 for Euclidean distance.