Hdsk.DecisionTree

Description

This module facilitates the definition of decision tree classifiers with tools for combining and parametrizing entropy measures, pruning tactics, and more.

Documentation

classify

Arguments

:: (Foldable f, Alternative f, MonadPlus f, Eq label, Ord label)

=> (f tup -> [[tup -> Bool]])
Function to generate possible branchings from the data

-> (tup -> label)
Function which extracts the label from a data object

-> ((tup -> label) -> f tup -> [tup -> Bool] -> Double)
Split criteria, like infoGain or gainRatio

-> f tup
The dataset as a list of data objects

-> tup
The unobserved instance

-> label
The predicted label

Predict the label of an unobserved tuple, according to the provided parameters.

mkCatTests :: (Foldable f, Alternative f, Ord a) => (tup -> a) -> f tup -> f (tup -> Bool)

O(n log n) Compute the branching predicates for the feature values at the given index. The feature values at index idx should be categorical.

mkContTests :: (Foldable f, Alternative f, Ord a) => (tup -> a) -> f tup -> f (tup -> Bool)

O(???) Compute the branching predicates for a continuous predicate. Will always be a pair, where one branch claims the objects with values less than or equal to the split point, and the other those greater. The split optimizes the given criterion.