Hdsk.DecisionTree.Entropy

Description

This module implements various measures of entropy and information gain in service of decision tree classification.

Documentation

entropy :: (Functor f, Foldable f, Eq label, Ord label, Floating a) => (tup -> label) -> f tup -> a

O(n log n) Calculate the total information entropy of a dataset.

conditionalEntropy :: Eq label => (tup -> label) -> [tup] -> [tup -> Bool] -> Double

O(nC) Compute the conditional entropy of splitting the dataset over the given branches.