Hdsk.Description

Description

This module contains basic numerical descriptive statistics that are applicable to a vector of doubles.

Documentation

mean :: Vector Double -> Double

O(n) Computes the arithmetic mean of a vector of doubles.

var :: Vector Double -> Double

O(n) Computes the unbiased variance of a vector of doubles.

std :: Vector Double -> Double

O(n) Computes the standard deviation of a vector of doubles.

percentile :: Double -> Vector Double -> Double

O(n log n) Finds the pth percentile of the vector of doubles, i.e. the element which is greater than p% of the other elements. Uses fractional interpolation when p does not generate a whole index.

Unfortunately, the equation for fractional interpolation breaks down both at the bounds of lists and for lists shorter than 2 elements. On the flip side, these special cases can be handled in O(n) or even O(1) time for the singleton case.

q1 :: Vector Double -> Double

O(n log n) Finds the first quartile of a vector of doubles.

q3 :: Vector Double -> Double

O(n log n) Finds the third quartile of a vector of doubles.

median :: Vector Double -> Double

O(n log n) Finds the median element the vector. Undefined for empty vectors.

iqr :: Vector Double -> Double

O(n log n) Finds the interquartile range of a vector of doubles.

noOutliers :: Vector Double -> Vector Double