Hdsk.Metrics

Description

A standard set of metrics for evaluating both classification and regression models. In Big-O notation, let C represent the number of classes and n the number of predictions. In general, we assume n >> C, so that n terms dominate C terms.

Documentation

accuracy :: (Eq a, Fractional b) => [a] -> [a] -> [a] -> b

O(n) Given the ground truth yTrue and predictions yPred, the expression accuracy c yTrue yPred reports the accuracy of the predictions, (TP + TN) / (P + N). Here, c is the set of class labels, e.g. S.fromList ["cat", "dog"]. The proportion of classifications which were correct.

accuracyCM :: Fractional a => Matrix Int -> a

O(C^2) where C is the number of classes. Calculate accuracy from a confusion matrix, rather than a list of truths and predictions.

precision :: (Eq a, Fractional b) => [a] -> a -> [a] -> [a] -> b

O(n) Compute the precision (positive predictive value) of a list of predictions, given the class list, target class label, and ground truths. The below example show how to find precision for a binary classifier.

>>> truth = ["cat", "cat", "not cat"]
>>> preds = ["cat", "not cat", "not cat"]
>>> precision ["cat", "not cat"] "cat" truth preds
0.5

The expression TP / (TP + FP) represents precision in terms of counts of true and false predictions. It is the proportion of positive predictions which are true.

Precision is undefined when no positive predictions are made.

precisionCM :: Fractional a => Matrix Int -> Int -> a

O(C^2) Compute the precision directly from a confusion matrix. The second argument is the class index within the confusion matrix. For instance, given the class list

>>> ["dog", "cat"]

The index of class dog is 1 and the index of class cat is 2 (since the matrix is 1-indexed).

recall :: (Eq a, Fractional b) => [a] -> a -> [a] -> [a] -> b

O(n) Compute the recall (sensitivity, true positive/ hit rate) of a list of predictions, given ground truth, class list, and target class label. See precision for discussion of arguments.

Recall, in terms of the confusion matrix, is TP / (TP + FN), and represents the proportion of positive predictions which were classified as such.

Recall is undefined when there are no positive observations.

recallCM :: Fractional a => Matrix Int -> Int -> a

O(C^2) Compute recall from a confusion matrix for a specified class. See precisionCM for discussion on the class index argument.

specificity :: (Eq a, Fractional b) => [a] -> a -> [a] -> [a] -> b

O(???) Compute the specificity (true negative rate) of the predictions given a list of class labels, a target class, and ground truth. TN / (FP + TN)

specificityCM :: Fractional a => Matrix Int -> Int -> a

f1 :: (Eq a, Fractional b) => [a] -> a -> [a] -> [a] -> b

O(???) Compute the f1-score of the model for a given class.

f1CM :: Fractional a => Matrix Int -> Int -> a

confusionMatrix :: Eq a => [a] -> [a] -> [a] -> Matrix Int

O(C^2 + n) where C is the number of classes and n the number of predictions. In general, n >> C.

Generates the confusion matrix for the predictions of an N-class predictor. The result will be a 1-indexed NxN matrix where rows represent the predicted class and columns the actual class. Classes are encoded as indices, where the index of a class within the matrix corresponds to its index within the set of classes.

tp :: Num a => Matrix Int -> Int -> a

O(1)

fp :: Num a => Matrix Int -> Int -> a

O(C)

tn :: Num a => Matrix Int -> Int -> a

O(C^2)

fn :: Num a => Matrix Int -> Int -> a

O(C)