Hdsk.NearestNeighbors

Description

This module implements the K Nearest Neighbors classifier, which finds the k tuples "closest" to the unobserved instance and combines their labels via some voting/ consensus metric to produce a prediction.

Documentation

knn :: (tup -> tup -> Double) -> Int -> [tup] -> tup -> [tup]

O(nD log k) where D is the cost of the distance function. Select the k nearest training instances to the argument tuple, according to the given distance metric (greater value means more dissimilar).

majorityLabel :: Eq label => (tup -> label) -> [tup] -> label

O(k^2) Report the most frequent label from a list of data instances.