Hdsk.Numerical

Description

This module contains functions which support efficient numerical methods, e.g. approximating derivatives.

Documentation

terminate :: Double -> (a -> Double) -> [a] -> a

Consume a stream of sequentially improving estimations until some measure of error fails to improve by eta. Requires at least two estimates to be produced by the stream. Assumes that error decreases with each estimate.