Hdsk.Preprocessing

Description

Hdsk.Preprocessing exposes a suite of data preprocessing utilities.

Documentation

newtype PreprocessedBy (algs :: [Type]) a

The PreprocessedBy type tags datasets of type a as having been preprocessed by algorithm alg. This is useful for requiring certain preprocessing steps to have occurred before modeling.

Constructors

Preprocessed a

data Normalizer

Algorithms which are Normalizers scale data in a way that can be said to eliminate units, enabling the comparison of features which, raw, have different scales or even orders of magnitude.

data MissingValueHandler

MissingValueHandlers denote that some consideration has been given to missing values within a dataset. This could be anything from dropping all rows with a missing value (as with TODO), to allowing all missing values to pass through (i.e. TODO).

dropMissingVals :: PreprocessedBy algs a -> PreprocessedBy (MissingValueHandler ': algs) (f a)

keepMissingVals :: PreprocessedBy algs a -> PreprocessedBy (MissingValueHandler ': algs) (f a)

standardScaler :: PreprocessedBy algs a -> PreprocessedBy (Normalizer ': algs) (f a)

Standardize features by removing the mean and scaling to unit variance.

minMaxScaler :: PreprocessedBy algs a -> PreprocessedBy (Normalizer ': algs) (f a)

Transform features by scaling each feature to a given range.

noPreprocess :: f a -> PreprocessedBy '[] (f a)