Hdsk.Training

Description

TODO

Documentation

data SplitData t a

Tagging and tracking partitioned data. A term of type SplitData t a refers to a partition of a dataset of type a tagged with annotation t, commonly either Train or Test.

data Train

A tag for SplitData which marks the training partition of a dataset.

data Test

A tag for SplitData which marks the testing partition of a dataset.

trainTestSplit :: PreprocessedBy '[] (f a) -> (SplitData Train a, SplitData Test a)

Split a dataset into a training and a testing partition. Note that the preprocessing stack must be empty; this prevents preprocessors from leaking information from the testing partition into model training.