Hdsk.Util

Description

This module contains several general utility functions used throughout the library. Exporting this because they may be helpful to users too.

Documentation

count :: (Foldable f, Ord k) => f k -> Map k Int

O(n log n) Given a container, construct a map from the unique elements of the container to their frequency within the container.

countHT :: (Foldable f, Eq k, Hashable k) => f k -> ST s (HashTable s k Int)

O(n) Given a container, construct a hash table from the unique elements of the container to their frequency within the container. I recommend this function over count when 1) counting performance is critical, AND 2) you're familiar with the ST monad.

countElemHT :: (Foldable f, Eq k, Hashable k) => f k -> k -> Int

O(n) Convenience function for getting the count of a single element from a container.