Hdsk.Bayesian.NaiveBayes

Description

A simple implementation of a Naive Bayes classifier for purely categorical data.

Documentation

type CPT = Map String (Map String Int)

Type alias for the conditional probability table data structure. Maps labels to lookup tables where lookup tables map from feature values to the number of times that feature value was observed with the parent label.

mkTables :: [[String]] -> [CPT]

/O(n * k * (log l + log m)) where n is the number of data points, k is the length of each data point, l is the number of labels, and m is the gretest number of feature values for any feature./ Generate the conditional probability tables over the dataset. Counts the occurences of each feature value given a class. Maps from class label to a map from feature value to count.

classify :: [CPT] -> [String] -> String

/O(l k m) where l is the number of labels, k is the length of each feature vector, and m is the number of feature values./ Given a list of conditional probability tables and an unlabeled tuple, predicts the label of the tuple.

posterior :: [CPT] -> [String] -> String -> Double

/O(k m) where k is the length of the feature vector and m is the number of feature values./ Given conditional probability tables, a tuple, and a label calculate the posterior probability of the label given that tuple. P(C|X)

likelihood :: String -> String -> CPT -> Double

O(m) Given a class, feature value, and conditional probability table, compute the likelihood of that configuration. P(x_i|c)

labelCount :: Num n => CPT -> String -> n

O(m) where m is the number of feature values. Computes the frequency of the given label from the given conditional prob. table.

dataLength :: Num n => CPT -> n

O(l m) Calculate the number of training instances from a CPT.