Hdsk.Cluster

Description

Here we implement a number of different clustering algorithms.

Documentation

dbscan :: Member algs Normalizer (Text "DBSCAN works best with normalized data." :$$: Text "Please apply a Normalizer to your data before moving on.") => PreprocessedBy algs (f a) -> f clusterId

Use DBSCAN to compute a clustering of the pre-normalized dataset.

kmeans :: Member algs Normalizer (Text "KMeans doesn't give meaningful results on unnormalized data." :$$: Text "Please apply a Normalizer to your data first.") => PreprocessedBy algs (f a) -> f clusterId

Calculate the clustering of the pre-normalized dataset using KMeans. TODO: check assumption of convex clusters, as well as choice of K?