Hdsk.DecisionTree

Description

This module facilitates the definition of decision tree classifiers with tools for combining and parametrizing entropy measures, pruning tactics, and more.

Documentation

classify

Arguments

:: (Foldable f, Alternative f, MonadPlus f, Eq label, Ord label)

=> (f tup -> [[tup -> Bool]])
Function to generate possible branchings from the data

-> (tup -> label)
Function which extracts the label from a data object

-> ((tup -> label) -> f tup -> [tup -> Bool] -> Double)
Split criteria, like infoGain or gainRatio

-> f tup
The dataset as a list of data objects

-> tup
The unobserved instance

-> label
The predicted label

Predict the label of an unobserved tuple, according to the provided parameters.

infoGain :: (Foldable f, Alternative f, Eq label, Ord label) => (tup -> label) -> f tup -> [tup -> Bool] -> Double

O(nC + n log n) Compute the information gain of the given branching. Branches are encoded as a list of predicates over data instances. An object belongs to whichever branch for which it passes the predicate. Assumes that each data object will pass one and only one of the branching predicates.

splitInfo :: (Foldable f, Alternative f, Eq label, Ord label) => (tup -> label) -> f tup -> [tup -> Bool] -> Double

O(nB) where B is the number of branches. Compute the split info of the branching.

gainRatio :: (Foldable f, Alternative f, Eq label, Ord label) => (tup -> label) -> f tup -> [tup -> Bool] -> Double

O(nB + nC + n log n) Compute the gain ratio of a branching.

mkCatTests :: (Foldable f, Alternative f, Ord a) => (tup -> a) -> f tup -> f (tup -> Bool)

O(n log n) Compute the branching predicates for the feature values at the given index. The feature values at index idx should be categorical.