Hdsk.DecisionTree

Description

This module facilitates the definition of decision tree classifiers with tools for combining and parametrizing entropy measures, pruning tactics, and more.

Documentation

data DecisionTree tup label

Representation of a decision tree's structure. Non-leaf nodes are encoded with a predicate over the tuple type (this function signals whether a tuple belongs to the branch) and its child nodes. Leaf nodes, or decisions, likewise claim tuples with their embedded predicate, and store the appropriate label.

Constructors

Branches (tup -> Bool) [DecisionTree tup label]
The first field of a branching node is the predicate which encodes it (e.g. "this is the branch for tuples with attribute weather = sunny), and the second is the node's children.

Decision (tup -> Bool) label
Again, the first field determines which data objects this node applies to, and the second records the decision of the node.

data Attribute tup v

An attribute is a value of a data object which is not its label; a data objects attributes are the complete set of non-label values which can be extracte from the data object. Each one can be encoded as a function from a data object to the target value type, and wrapping that function in the Attribute type allows you to tag which attributes should be treated as continuous, and which are categorical (a distinction which must be made for certain algorithms).

Constructors

Categorical (tup -> v)
In general, categorical variables are those which can only take on one of a certain, fixed set of discrete values. In the context of this data type, an attribute should be tagged as categorical if its branches should be encoded with equality. That is, when a tree branches on a Categorical attribute, it will generate one branch for each of the feature values, and each will test whether an unobserved data object shares that feature value.

Ordinal (tup -> v)
In general, ordinal data are values which can be meaningfully ordered relative to one another. This can be either discrete (e.g. military rank, grade in school) or continuous (temperature outside). In the context of this data type, attributes should be tagged as Ordinal when the branches they generate should be encoded as a binary split over some pivot. That is, when a tree branches on an Ordinal attribute, it generates two branches, one which accepts data objects with attribute vaules less than or equal to a certain pivot, and one for those with greater (the pivot value depends on the split criteria for the algorithm).

classify :: DecisionTree tup label -> tup -> label

O(X log D) where X is the maximum number of feature values and D is the depth of the tree. Use the given decision tree to predict a label for an unobserved data instance.

id3 :: forall f tup label v. (Alternative f, MonadPlus f, Foldable f, Ord label, Ord v) => label -> (tup -> Bool) -> (tup -> label) -> [Attribute tup v] -> f tup -> DecisionTree tup label

O(???) Generate a decision tree using the ID3 algorithm.