Hdsk.Description

Description

This module contains basic numerical descriptive statistics that are applicable to a vector of doubles.

Documentation

mean :: (Foldable f, Fractional n) => f n -> n

O(n) Computes the arithmetic mean of a collection of numbers.

var :: (Foldable f, Functor f, Floating n) => f n -> n

O(n) Computes the unbiased variance of a collection of numbers.

std :: (Foldable f, Functor f, Floating n) => f n -> n

O(n) Computes the standard deviation of a collection of numbers.

percentile :: (Selectable p, RealFrac n) => n -> p n -> n

O(n) Selects the element which is greater than p% of the rest. When the p-th percentile does not land directly on a whole index, midpoint interpolation is used to average left and right side of the split.

median :: (Selectable p, RealFrac n) => p n -> n

O(n) Finds the median element the collection.

q1 :: (Selectable p, RealFrac n) => p n -> n

O(n) Finds the first quartile of a collection of numbers.

q3 :: (Selectable p, RealFrac n) => p n -> n

O(n) Finds the third quartile of a collection of numbers.

iqr :: (Selectable p, RealFrac n) => p n -> n

O(n) Inter-quartile range. The distance between the first and third quartiles.

select :: (Selectable p, Ord a) => Int -> p a -> a

O(n) Simple implementation of quickselct (aka Hoare's algorithm or k-rank). Selects the k-smallest element from the collection.

class Foldable p => Selectable p

Defines a container which is suitable for the k-rank/ select algorithm.

Minimal complete definition

partition, head, tail