Hdsk.Description

Description

This module contains basic numerical descriptive statistics that are applicable to a list of numbers.

Documentation

mean :: (Real r, Fractional p) => [r] -> p

Computes the arithmetic mean of a list of real numbers.

>>> mean [1, 3, 3, 4]
2.75
>>> mean [10]
10.0

var :: (Real p, Floating p) => [p] -> p

Computes the variance of a list of numbers.

>>> var [1, 2]
0.25
>>> var [10.12, 10.12]
0.0

std :: (Real p, Floating p) => [p] -> p

Computes the standard deviation of a list of numbers. Examples:

>>> std [1, 1, 1]
0.0
>>> std [1, 2]
0.5