Hdsk.Util

Description

This module contains several general utility functions used throughout the library. Exporting this because they may be helpful to users too.

Documentation

majorityLabel :: (Foldable f, Functor f, Eq label, Ord label) => (tup -> label) -> f tup -> label

O(n log n) Compute the most frequent label in the data set. Ties do not have well defined behavior; the tie is broken by maximumBy which doesn't have clear documentation on ties.

count :: (Foldable f, Ord k) => f k -> Map k Int

O(n log n) Given a container, construct a map from the unique elements of the container to their frequency within the container.

countBy :: Foldable f => (a -> Bool) -> f a -> Int

O(n) Given a container and a predicate, calculate the number of elements which pass the predicate

countBy' :: (Foldable f, Num b) => (a -> Bool) -> f a -> b

O(n) Version of countBy with generic result.

countHT :: (Foldable f, Eq k, Hashable k) => f k -> ST s (HashTable s k Int)

O(n) Given a container, construct a hash table from the unique elements of the container to their frequency within the container. I recommend this function over count when 1) counting performance is critical, AND 2) you're familiar with the ST monad.

countElemHT :: (Foldable f, Eq k, Hashable k) => f k -> k -> Int

O(n) Convenience function for getting the count of a single element from a container.

length' :: (Num a, Foldable f) => f b -> a

O(n) Shorthand for genericLength which works on all foldables.

lg :: Floating a => a -> a

O(1) Shorthand for logBase 2

uniq :: Ord a => [a] -> [a]

O(n log n) Filter the duplicates out of a list.

uniq' :: (Foldable f, Alternative f, Ord a) => f a -> f a

O(n log n + nX) where X is the const of the associative binary operator for the given Alternative type. Filter the duplicates out of a Foldable, Alternative container.

nuniq :: (Foldable f, Ord a) => f a -> Int

O(n) Return the number of uniq values in the container.

head' :: Foldable f => f a -> a

O(n) Return the first item in a foldable.

last' :: Foldable f => f a -> a

O(n) Return the last element of a foldable.

listMap :: Foldable f => (a -> b) -> f a -> [b]

O(n) Map a function over a foldable, producing a list.